implement Daubechies wavelet in the LFSR.
A Daubechies wavelet is a type of orthogonal wavelet that has a maximal number of vanishing moments for some given support12. It is based on the work of Ingrid Daubechies, who invented compactly supported orthonormal wavelets3. They are used in signal denoising, compression, and Galerkin methods for PDEs45.

To implement a Daubechies wavelet using the above code, you need to modify the tap positions and feedback masks according to the filter coefficients of the desired wavelet. For example, if you want to use a Daubechies 4 (db4) wavelet, which has four vanishing moments and eight filter coefficients4, you can use these values:

Note that it was designed for different LFSR length.